libc/startup: Some more reasonable startup code for RISCovite
This is still not complete but should at least be enough to find the
arguments, environment variables, and aux vector passed by the supervisor
on entry.

We still have more work to do with threading support before we can complete
this by setting up the initial thread's TLS area and associated data
structures.