compiler-rt: Hackily opt RISCovite out of some OS-specific builtins
This history around these three special exceptions is hard to track down:
I have seen comment that eprintf.c is really only for Apple platforms, but
the reasons for the others being special-cased remain unclear.

In the long run we should find a way to exclude these on RISCovite
properly, but it's not clear to me yet where this FUSCHIA symbol is being
defined and so we don't have an equivalent one for RISCOVITE, and even if
we did it feels to me like these ought to be _opt-in_ for the specific
platforms that need them, rather than each new platform having to opt out
of including them.