libc: Stub of malloc/etc for RISCovite
For the moment we're interested only in bringing this up enough to get
the toolchain to build cleanly, so this adds a malloc that just always
fails to give the RISCovite startup code something to link to, even though
of course it won't actually work at runtime until we have a real allocator
here.

At this point it's not clear exactly what strategy we'll take for a real
memory allocator. There are two in the LLVM tree already: scudo and
rtmalloc. The former is quite complex because it's designed to be
"hardened". The latter seems more appropriate but there doesn't seem to be
any existing support for using it with LLVM libc -- it's primarily there
to support the toolchain itself on platforms that have an inefficient
system-provided malloc.