libc: stdlib atexit depends on libc.src.__support.CPP.new
The build scripts apparently cannot discover this indirect dependency
automatically. It seems that the CMake function that discovers all of the
needed objects doesn't expect that a header dependency might lead to an
object dependency.

Therefore we add an artificial extra dependency directly from atexit to
libc.src.__support.CPP.new, thereby causing its object (and anything else
it depends on) to be included in libc.a.