libc: RISCovite pthread-related stubbing
The pthread implementation and its associated helpers are currently making
a bunch of Linux-specific assumptions without explicit abstraction, and so
getting this wired in for RISCovite was a little tricky.

This doesn't actually propertly implement the pthread API yet. Instead,
this is just enough to make it compile _at all_ with the relevant helper
functions ultimately just stubbed to always report failure.

A real implementation of the pthread API (in userspace, using timer
interrupts) will follow in later commits.