libc: More work on RISCovite multithreading (still not complete)
This commit is a further increment on stubbing out the userspace
multithreading implementation. This gets us enough to have the first
call to pthread_create allocate the additional resources we need for our
threading runtime, including some additional memory to track the overall
state of the system and a software interrupt object that we'll use to
force thread switching when a running thread puts itself to sleep while
waiting for something.

This also includes some helper glue code for turning a supervisor-provided
interrupt stack frame into a TaskState object and then delegating to a
function that receives a pointer to that object, which we'll use for both
the forced task switching and preemption interrupt handlers to make sure
we can always accurately save the entire state of the current thread.