libc: RISCovite userspace threading implementation continues
This is just enough to get a thread started and do some preemptive thread
switching between the runnable threads.

However, as of this commit a non-main thread just stalls in a wfi loop
once its exits, because we don't yet have anywhere to track thread results
nor code to tear down a thread once it is no longer needed.