libc: RISCovite userspace thread cleanup
Rather than having a thread try to clean itself up while running on its
own stack (and thereby having to be careful not to use the stack it's
already freed) we'll instead maintain a list of "zombie" threads and use
a separate cleanup thread to free them.

Implementing this as a special thread instead of just inline in the
thread switching function means that we can do this work at interrupt
priority floor zero and thus minimize delay to any other interrupt
handlers that might need to run.

However, the cleanup thread is effectively given an unlimited timeslice to
do its work and so no other thread can be scheduled again until the zombie
list is empty. That compromise means that the cleanup thread can assume it
has exclusive control over the zombie list and thus it does not need to
mask interrupts while doing any of its work.