libc: RISCovite mutex support
This is some admittedly-somewhat-clunky support for mutexes in the
userspace threads implementation used by RISCovite's libc.

Various platform-agnostic parts of libc, such as stdio streams and the
atexit mechanism, rely on this mutex API. Implementing this therefore makes
it safe to use those mechanisms from multiple threads concurrently.

This implementation is not very well optimized yet since my priority is
just to get a proof-of-concept implementation. It remains to be seen
whether RISCovite programs will make much use of libc threads in practice;
it's implemented here mainly just to ease the work of porting software
that was originally written for other platforms.